Move constants, add sources
